requirements.txt
streamlit==1.31.0
transformers==4.36.0
torch==2.1.0
sentencepiece==0.1.99
sacremoses==0.1.1
Pillow==10.1.0
pytesseract==0.3.10
PyPDF2==3.0.1
python-docx==1.1.0
SpeechRecognition==3.10.1
pydub==0.25.1
protobuf==3.20.3
